Adding the main canvasGame class object with all it's methods and functionality.
This is very early spaghetti code and will obviously require a lot of refactoring going forward, but for now it serves as a good idea of what's possible visually.  Note that almost everything is focused on rendering right now and there's next to zero "game" functionality.  Just a bunch of sprite tests.